Ka 228 search result screen

Approved-by: Vivek Pandey

+++ app/src/main/java/agstack/gramophone/ui/farm/model/AddHarvestRequest.kt
@@ -2,6 +2,6 @@ package agstack.gramophone.ui.farm.model
 
 data class AddHarvestRequest(
     val farm_reference_id: String,
-    val output_qty: Int,
+    val output_qty: Double,
     val output_unit_id: String
 )
